Add lower upper case custom validator to the password field and moved the components to a separate folder
