+++ src/app/app-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
-import { SignUpComponent } from './sign-up/sign-up.component';
+import { PageNotFoundComponent } from './components/page-not-found/page-not-found.component';
+import { SignUpComponent } from './components/sign-up/sign-up.component';
 
 const routes: Routes = [
   { path: 'sign-up', component: SignUpComponent },
