+++ src/app/components/sign-up/sign-up.component.sass
