--- src/app/sign-up/sign-up.component.html
@@ -1,60 +1,0 @@
-<div class="sign-up-container mx-auto">
-  <div class="card">
-    <div class="card-header">Sign up</div>
-    <div class="card-body">
-      <div class="input-group mb-2 mr-sm-2">
-        <div class="input-group-prepend">
-          <div class="input-group-text">
-            <i class="bi bi-person-circle"></i>
-          </div>
-        </div>
-        <input
-          type="text"
-          class="form-control"
-          id="firstName"
-          placeholder="First name"
-        />
-      </div>
-      <div class="input-group mb-2 mr-sm-2">
-        <div class="input-group-prepend">
-          <div class="input-group-text">
-            <i class="bi bi-people-fill"></i>
-          </div>
-        </div>
-        <input
-          type="text"
-          class="form-control"
-          id="lastName"
-          placeholder="Last name"
-        />
-      </div>
-      <div class="input-group mb-2 mr-sm-2">
-        <div class="input-group-prepend">
-          <div class="input-group-text">
-            <i class="bi bi-mailbox2"></i>
-          </div>
-        </div>
-        <input
-          type="email"
-          class="form-control"
-          id="email"
-          placeholder="Email"
-        />
-      </div>
-      <div class="input-group mb-2 mr-sm-2">
-        <div class="input-group-prepend">
-          <div class="input-group-text">
-            <i class="bi bi-lock-fill"></i>
-          </div>
-        </div>
-        <input
-          type="password"
-          class="form-control"
-          id="password"
-          placeholder="Password"
-        />
-      </div>
-      <button class="btn btn-primary d-flex ml-auto">Register</button>
-    </div>
-  </div>
-</div>
